Allow restricting speech to a particular region of the screen
This can be useful to prevent unwanted things from being spoken. For instance,
recent versions of vim appear to briefly print the last key entered in the
lower right corner of the screen, even if noruler is set. Add a "speaking
window" function that can be bound to a key. This takes two forms:
- With no arguments, clears the region so that all text is spoken.
- With four arguments, specifies the top/left/bottom/right bounds of the
region in which text is to be spoken. Row/column indices are 1-based. A
value of 0 indicates the edge of the screen. A negative value can be used
to specify an offset from the bottom or right edge of the screen.

For example, if the following is added into the [normalkeys] section in
yasr.conf:

0x1b60:speaking window
0x1b31:speaking window:1:1:-1:0

then two key bindings are added:
alt-` will clear the existing screen region.
alt-1 will set the screen region to exclude the last line of the screen.

Also remove the "special" option. This was a hack to allow only the top
line of the screen to be read. This can now be done by setting a region.