Add a "silent window" key binding
This allows the inverse of a speaking window. If a region is set, then
speech within the screen region is not spoken. For instance, for a 75x240
terminal,
0x1b31:silent window:75:229:75:240
causes alt-1 to set a silent region for the last 12 characters of the last
line. (Trying to use negative values here would likely not work well; the
semantics would need to be better defined.)